+++ Code/LibMain.lua
@@ -33,20 +33,21 @@ For more information on this, and how to apply and follow the GNU AGPL, see
 --]]
 
 --logging variables
-FFL_Debugging = false
 FFL_Log = { "*** Fizzle Fuze's Library loaded! *** "}
+if not SharedModEnv["FFL_Debug"] then
+    SharedModEnv["FFL_Debug"] = false
+end
 
 --print log messages to console and disk
 local function FFL_PrintLog()
 
-    local MsgLog = FFL_Log
-    if #MsgLog > 0 then
-        for _, Msg in ipairs(MsgLog) do
+    if #FFL_Log > 0 then
+        for _, Msg in ipairs(FFL_Log) do
             print(Msg)
         end
         FlushLogFile()
 
-        MsgLog = {}
+        FFL_Log = {}
         return
     end
 end
@@ -55,12 +56,11 @@ end
 function FFL_LogMessage(...)
     local Sev, Arg = nil, {...}
     local SevType = {"INFO", "DEBUG", "WARNING", "ERROR", "CRITICAL"}
-    local MsgLog = FFL_Log
 
     if #Arg == 0 then
         print("/?.lua CRITICAL: No error message!")
         FlushLogFile()
-        MsgLog[#MsgLog+1] = "/?.lua CRITICAL: No error message!"
+        FFL_Log[#FFL_Log+1] = "/?.lua CRITICAL: No error message!"
         return
     end
 
@@ -77,17 +77,17 @@ function FFL_LogMessage(...)
         Arg[3] = "DEBUG: "..Arg[3]
     end
 
-    if Sev == "DEBUG: " or Sev == "INFO: " and not FFL_Debugging then
+    if (Sev == "DEBUG: " or Sev == "INFO: ") and not SharedModEnv["FFL_Debug"] then
         return
     end
 
     local Msg = Arg[1].."/"..Arg[2]..".lua "
     for i = 3, #Arg do
-        Msg = Msg..tostring(Arg[i])
+        Msg = Msg .. tostring(Arg[i])
     end
-    MsgLog[#MsgLog+1] = Msg
+    FFL_Log[#FFL_Log + 1] = Msg
 
-    if FFL_Debugging then
+    if SharedModEnv["FFL_Debug"] then
         FFL_PrintLog()
     end
 end
@@ -101,7 +101,7 @@ function FFL_Translate(TString)
     local id = IsT(TString)
 
     if type(id) == "number" then -- already has a translation
-        return id, TString
+        return T(TString)
     else
         id = RandomLocId()
         if type(id) ~= "number" then
@@ -111,12 +111,26 @@ function FFL_Translate(TString)
     end
 end
 
-function OnMsg.NewHour()
-    if FFL_Debugging == true then
+function FFL_FormatNumber(Number, AsT)
+    AsT = AsT or false
+
+    local Ret = InfobarObj.FmtRes(nil, Number)
+    if not AsT then
+        Ret = _InternalTranslate(Ret)
+    end
+    return Ret
+end
+
+function OnMsg.NewHour(Hour)
+    if SharedModEnv["FFL_Debug"] then
+        Log("New Hour: ", Hour)
         FFL_PrintLog()
     end
 end
 
-function OnMsg.NewDay()
+function OnMsg.NewDay(Day)
+    if SharedModEnv["FFL_Debug"] then
+        Log("New Day: ", Day)
+    end
     FFL_PrintLog()
 end