+++ changelog.md
@@ -1,5 +1,7 @@
 Version 1.4
+- Added number formation function
 - Updated translation functions
+- Updated logging functions
 
 Version 1.3
 - Another bugfix